refactor: move non-CacheHandler related types and utils from cache-types to blob-types module
