+++ src/shared/cache-types.cts
@@ -1,5 +1,3 @@
-import { isPromise } from 'node:util/types'
-
 import type {
   CacheHandler,
   CacheHandlerValue,
@@ -11,8 +9,6 @@ import type {
   IncrementalCachedPageValue,
   IncrementalCacheValue,
 } from 'next/dist/server/response-cache/types.js'
-
-import { recordWarning } from '../run/handlers/tracer.cjs'
 
 export type { CacheHandlerContext } from 'next/dist/server/lib/incremental-cache/index.js'
 
@@ -157,61 +153,4 @@ export type CacheHandlerForMultipleVersions = BaseCacheHandlerForMultipleVersion
     value: Parameters<BaseCacheHandlerForMultipleVersions['set']>[1],
     context: CacheHandlerSetContextForMultipleVersions,
   ) => ReturnType<BaseCacheHandlerForMultipleVersions['set']>
-}
-
-export type TagManifest = { revalidatedAt: number }
-
-export type HtmlBlob = {
-  html: string
-  isFallback: boolean
-}
-
-export type BlobType = NetlifyCacheHandlerValue | TagManifest | HtmlBlob
-
-const isTagManifest = (value: BlobType): value is TagManifest => {
-  return false
-}
-
-const isHtmlBlob = (value: BlobType): value is HtmlBlob => {
-  return false
-}
-
-export const estimateBlobSize = (valueToStore: BlobType | null | Promise<unknown>): number => {
-  // very approximate size calculation to avoid expensive exact size calculation
-  // inspired by https://github.com/vercel/next.js/blob/ed10f7ed0246fcc763194197eb9beebcbd063162/packages/next/src/server/lib/incremental-cache/file-system-cache.ts#L60-L79
-  if (valueToStore === null || isPromise(valueToStore) || isTagManifest(valueToStore)) {
-    return 25
-  }
-  if (isHtmlBlob(valueToStore)) {
-    return valueToStore.html.length
-  }
-  let knownKindFailed = false
-  try {
-    if (valueToStore.value?.kind === 'FETCH') {
-      return valueToStore.value.data.body.length
-    }
-    if (valueToStore.value?.kind === 'APP_PAGE') {
-      return valueToStore.value.html.length + (valueToStore.value.rscData?.length ?? 0)
-    }
-    if (valueToStore.value?.kind === 'PAGE' || valueToStore.value?.kind === 'PAGES') {
-      return valueToStore.value.html.length + JSON.stringify(valueToStore.value.pageData).length
-    }
-    if (valueToStore.value?.kind === 'ROUTE' || valueToStore.value?.kind === 'APP_ROUTE') {
-      return valueToStore.value.body.length
-    }
-  } catch {
-    // size calculation rely on the shape of the value, so if it's not what we expect, we fallback to JSON.stringify
-    knownKindFailed = true
-  }
-
-  // fallback for not known kinds or known kinds that did fail to calculate size
-  // we should also monitor cases when fallback is used because it's not the most efficient way to calculate/estimate size
-  // and might indicate need to make adjustments or additions to the size calculation
-  recordWarning(
-    new Error(
-      `Blob size calculation did fallback to JSON.stringify. Kind: KnownKindFailed: ${knownKindFailed}, ${valueToStore.value?.kind ?? 'undefined'}`,
-    ),
-  )
-
-  return JSON.stringify(valueToStore).length
 }
